requirements.txt
streamlit
pandas
scikit-learn
plotly
watchdog
umap-learn
seaborn
statsmodels
matplotlib-venn
holoviews==1.15.4
bokeh==2.4.3
networkx
setuptools
wheel
numpy
numba>=0.55
llvmlite>=0.39.0
bs4
cProfile
pstats